Update data import
FGDinterface - add error trap if user cancels in getGrid. Bug fixes to addGrid
GDinterface - do not assume rows are time. allow user to pass prompt text and if not duration for rows prompt user to amend Dsproperty for Row. correct assignment of cline in addGrid
GD_ImportData - modified to be able to load a grid 'z' array from a mat file and add X and Y dimensions. Corrected addData to use updates to FGDinterface and GDinterface. Added setChannel method to handle additional channel properties in both loadData and addData.