Added FGD_ImportData and modified GD_ImportData to use GDinterface
Added FGD_ImportData for inlet specific grids. ChannelForm modified to use FDG_ImportData, WaveRayModel and EstuaryDB use GD_ImportData and ModelSkill modified to use both. Updated functions and classes that refer to either of these classes and tested WaveRayModel and ChannelForm to ensure old models still work. ModelSkill still needs testing. Note: tools in FGDinterface assume channel is along x-axis. This is not the case in EstuaryDB which is why it uses GD_ImportData and the Gdinterface.